Enhance anchoring by introducing an alternative way to specify the column
Anchoring two comments at the same column, when the comment keywords are
subsets or supersets of each other, will fail, because the shorter comment
keyword will always win. This is because the keywords are searched in
ascending order.

  Example:

  comment -
  comment --
  anchor 1 - --
  font a -
  font b --

  All lines starting with -- will be in font a, too!

To solve this problem, there is an alternative way to specify the column now.
A column number prefixed by a ~ character will be preferred to the same
column number without prefix.

  Example:

  comment -
  comment --
  anchor 1 -
  anchor ~1 --
  font a -
  font b --

  All lines starting with -- will be in font b now!

As a drawback, the maximum column number that can be used diminishes from
4094 (0xFFE) to 2046 (0x7FE).