Fix backslash handling
Even if not every \ at the beginning should result in a literal \
(but only in the case of \\, so that \( won't result in \(, but (,
and \$ won't result in \ and a variable substitution, but $), any
subsequent \\ must be treated as a quoted special character, namely \,
again.

(This condition has been changed from the previous version, where
subsequent \\ worked. As a side effect of the change, handling of
subsequent \\ has been broken.)