Fix nottyitalic
If it is unfortunate to mix colors and italics because the terminal
can't properly display it, nottyitalic shall allow to see (at least)
the colors on the screen. This works if the terminal implements what
Elvis uses to display italic. If the terminal, however, doesn't
implement what Elvis uses to display italic, underlining is selected
to represent italics - which gives a different result.

The behaviour for nottyitalic should be identical for both types of
terminals.

So move up the condition checking for nottyitalic so that COLOR_ITALIC
can get removed first of all - before it possibly will be replaced by
COLOR_UNDERLINED.