Improve handling of internal cut buffers with marks
Usually the nine internal "1 to "9 cut buffers get shifted so that the
oldest one gets lost and the last nine onces are still accessible.
Because cut buffers with marks can't be deleted, a problem arises if all
of them have marks.

By now the first cut buffer "1 will always be reused then (which
according to a comment in the source file is save), but the user will
always lose their last but one buffer information while "2 to "9 keep
sticking at their respective content.

So, even if all cut buffers have marks they still should shift as they
would do without marks. In order to accomplish that, "9 will temporarily
be renamed "~ and after the shifting renamed "1. That way, the reused
one is actually "9 while "1 to "8 have shifted nicely.