Fix bug with syntax coloring and anchor ^
Anchor ^ is used "to indicate that the keyword can only appear at the
front of a line (after optional whitespace but no printed characters)".

If a such anchored keyword is preceded by nothing but other keywords
and whitespace, it is handled as if nothing had been appeared before it.

Example:

  if gui == "termcap"

Although the equal sign and the quotation mark are defined "anchor ^",
they will be colored as keyword resp. comment.

Fix it by noting the end of indentation when a keyword has been
detected.