Improve bottom row coloring of windows without input focus
The bottom row of a window that currently doesn't have the input focus
is colored in the idle color. The ruler and the label for the current
parse state, however, are unaffected and retain their colors.

In order to have the bottom row completely colored in the idle color,
change the colors of the ruler and the label according to whether the
window has input focus or not.