Don't always move the cursor to the bottom of the screen on termcap end
This is only necessary if the terminal can't restore the screen as it
was before the invocation of the editor (and thus the content of the
edited file would remain on the screen).

The xterm terminal emulator with the :te: termcap entry enabled uses an
alternate screen and will perfectly restore the screen on endtcap().

This patch will prevent the numerous ugly empty lines between the prompt
where the editor had been called from and the prompt after the editor
has ended.