Fix segmentation fault with column based anchoring
Commit a3795a5728a6c2b9 introduced an alternative way to specify the
column for an anchor.

While searching for a node for an alternative anchor the reference
pointer to the end of the scanned text will move further, and if there
is no such node the reference pointer belonging to the node for the
normal anchor will be lost.

Save the reference pointer belonging to the node for the normal anchor
and restore it when returning this very node.

This closes issue #16 reported by fbrau on GitHub.